[CHK-702] fix: amount and language filter bug (#23)

+++ src/main/java/it/pagopa/ecommerce/payment/methods/infrastructure/rule/AmountTypeLangFilterRule.java
@@ -5,6 +5,8 @@ import it.pagopa.ecommerce.payment.methods.infrastructure.PspRepository;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
+import java.math.BigInteger;
+
 @Component
 public class AmountTypeLangFilterRule implements IFilterRule {
     @Override
@@ -14,6 +16,6 @@ public class AmountTypeLangFilterRule implements IFilterRule {
 
     @Override
     public Flux<PspDocument> execute(PspRepository pspRepository, Integer amount, String language, String paymentTypeCode) {
-        return pspRepository.findPspMatchAmountTypeLang(amount, paymentTypeCode, language);
+        return pspRepository.findPspMatchAmountTypeLang(BigInteger.valueOf(amount), paymentTypeCode, language);
     }
 }
